Make prox trigger relative, not absolute
The screen turn on point of proximity >= 5000 was not working for some users; it turns out there is a wide variability in the absolute proximity values reported by different Wink Relays (with some having a value at rest > 5000). This commit changes the turn-on point to a value relative to the last reading of the prox sensor, with the default turn-on point set at 0.75% higher than the previous reading. Two new variables were introduced: "prox_delta" (a config parameter) that sets the percentage change value and "delta" which is the local variable in main() that takes the default or user-set value of prox_delta. Added info re: setting prox sensor sentitivty in README.md.